+++ dialect/mysql/dialect/mods.go
@@ -2,7 +2,6 @@ package dialect
 
 import (
 	"context"
-	"fmt"
 	"io"
 
 	"github.com/stephenafamo/bob"
@@ -261,17 +260,6 @@ func (j JoinChain[Q]) Using(using ...string) bob.Mod[Q] {
 	return mods.Join[Q](jo)
 }
 
-type collation struct {
-	name string
-}
-
-func (c collation) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, _ int) ([]any, error) {
-	if _, err := fmt.Fprintf(w, " COLLATE %s", c.name); err != nil {
-		return nil, err
-	}
-	return nil, nil
-}
-
 type OrderBy[Q interface{ AppendOrder(clause.OrderDef) }] func() clause.OrderDef
 
 func (s OrderBy[Q]) Apply(q Q) {
@@ -298,7 +286,7 @@ func (o OrderBy[Q]) Desc() OrderBy[Q] {
 
 func (o OrderBy[Q]) Collate(collationName string) OrderBy[Q] {
 	order := o()
-	order.Collation = collation{name: collationName}
+	order.Collation = collationName
 
 	return OrderBy[Q](func() clause.OrderDef {
 		return order