+++ dialect/psql/dialect/mods.go
@@ -219,18 +219,6 @@ func (j CrossJoinChain[Q]) As(alias string, columns ...string) bob.Mod[Q] {
 	})
 }
 
-type collation struct {
-	name string
-}
-
-func (c collation) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, _ int) ([]any, error) {
-	if _, err := w.Write([]byte(" COLLATE ")); err != nil {
-		return nil, err
-	}
-	d.WriteQuoted(w, c.name)
-	return nil, nil
-}
-
 type OrderBy[Q interface{ AppendOrder(clause.OrderDef) }] func() clause.OrderDef
 
 func (s OrderBy[Q]) Apply(q Q) {
@@ -284,7 +272,7 @@ func (o OrderBy[Q]) NullsLast() OrderBy[Q] {
 
 func (o OrderBy[Q]) Collate(collationName string) OrderBy[Q] {
 	order := o()
-	order.Collation = collation{name: collationName}
+	order.Collation = collationName
 
 	return OrderBy[Q](func() clause.OrderDef {
 		return order